Split the CLI implementation file
