+++ lib/plugins/inspec-init/test/functional/inspec_init_profile_test.rb
